added documentation to components and functions. moved default tops and piece data into separate file.
