+++ client/src/components/DetailsBar.jsx
@@ -3,6 +3,9 @@ import React, {useState, useContext} from 'react';
 import Filters from './Filters.jsx';
 import GarmentContext from './../context.js';
 
+//Details Bar:
+  //Allows user to set boundries, neckline and filters for top base and detail
+  //Allows user to save tops
 function DetailsBar(props) {
   const {savedTops, setSavedTops} = props;
   const {top, setTop} = useContext(GarmentContext);
@@ -11,17 +14,27 @@ function DetailsBar(props) {
     neckline: ['', 'diamond-neck', 'deep-thin-neck'],
     lowerBoundary: ['', 'crop-top ', 'natural-waist']
   }
+  //adds current top to savedTops array
   const saveCurrentTop = () => {
     var topsCopy = savedTops.slice();
     topsCopy.push(top);
     setSavedTops(topsCopy)
   }
+  //updates top's filter (for the base and decoration so far)
+    //inputs:
+    //e: event from onChange
+    //type: 'hue', 'con', 'sat', 'op', 'sat'
+    //part: any key of the 'top' object that has the listed properties above
   var handleFilterChange = (e, type, part) => {
     var copy = Object.assign({}, top);
     copy[part][type] = e.target.value;
     setTop(copy);
-     console.log(top)
   }
+
+  //updates top's boundaries
+  //inputs:
+    //e: event from onChange
+    //type: string, name of type of boundary
   var handleBoundaryChange = (e, type) => {
     var index = e.target.value;
     var copy = Object.assign({}, top);
@@ -32,7 +45,6 @@ function DetailsBar(props) {
   return (
     <div className="details-bar">
       <h4>Details</h4>
-
       <label htmlFor="neckline">Neckline</label>
       <input type="range" name="neckline" min="0" max={neckline.length -1} step="1" onChange={(e) => {handleBoundaryChange(e, 'neckline')}}/>
       <label htmlFor="upper" >Upper Boundary</label>
