+++ client/src/components/Item.jsx
@@ -1,13 +1,15 @@
 import React, {useState, useContext} from 'react';
 
 //   white: {brightness: 3.5, saturation: 1, hue: 0, opacity: 100, contrast: 144.5},
-
+//creates string for filter css style from inputs
+//inputs: numbers for value of hue/saturation/briteness/contrast/opacity
 const makeFilterStyle = (hue, sat, brit, con, op) => {
   if (hue === undefined || sat === undefined || brit === undefined || con === undefined ||op === undefined) {
     return "brightness(" + 3.5 + ") hue-rotate(" + 0 + 'deg) saturate(' + 1 + '%) contrast(' + 144.5 + '%) opacity(' + 100 + '%)';
   }
   return "brightness(" + brit + ") hue-rotate(" + hue + 'deg) saturate(' + sat + '%) contrast(' + con + '%) opacity(' + op + '%)';
 }
+//Item: accepts data about garment/garment pieces and generates array of <img> elements
 function Item(props) {
   let {base, seamUrls, decoration, upperBoundary, neckline, lowerBoundary, zCounter, baseFilter, decorationFilter, displayClass, onClickEvent} = props;
 
@@ -26,13 +28,7 @@ function Item(props) {
   } else {
     decoration = {hasBoundaries: false};
   }
-  // const Decoration = (bool) => {
-  //   if (decoration) {
-  //     if (decoration.hasBoundaries === bool) {
-  //       return decorationsArr;
-  //     }
-  //   }
-  // }
+
   //base
   var baseImgs = null;
   if (base) {
